+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## v0.8.0
+
+- Deprecate `keyProp` in favor of new `identify` method
+- Optimized `weld`
+
 ## v0.7.1
 
 - Fixed `classes` edge cases